+++ src/Catalog/Product.php
@@ -100,6 +100,9 @@ final class Product
     /** @var \DateTimeImmutable|null */
     private $updatedAt;
 
+    /** @var \DateTimeImmutable|null */
+    private $availableSince;
+
     /**
      * @var Image[]
      */
@@ -164,6 +167,7 @@ final class Product
         $this->seoKeywords = $data['seoData']['keywords'] ?? '';
         $this->createdAt = isset($data['createdAt']) ? \DateTimeImmutable::createFromFormat(\DateTime::RFC3339, $data['createdAt']) : null;
         $this->updatedAt = isset($data['updatedAt']) ? \DateTimeImmutable::createFromFormat(\DateTime::RFC3339, $data['updatedAt']) : null;
+        $this->availableSince = isset($data['availableSince']) ? \DateTimeImmutable::createFromFormat(\DateTime::RFC3339, $data['availableSince']) : null;
         $this->infiniteStock = $data['infiniteStock'];
 
         if (!isset($data['images'])) {
@@ -464,6 +468,11 @@ final class Product
     public function getUpdatedAt(): ?\DateTimeInterface
     {
         return $this->updatedAt;
+    }
+
+    public function getAvailableSince(): ?\DateTimeInterface
+    {
+        return $this->availableSince;
     }
 
     /**