+++ tests/Catalog/CatalogServiceTest.php
@@ -110,7 +110,7 @@ final class CatalogServiceTest extends ApiTestCase
         $this->assertTrue($product->isTransactional());
         $this->assertSame(0.0, $product->getGreenTax());
         $this->assertSame(1.23, $product->getWeight());
-        $this->assertEquals(0, $product->getAverageRating());
+        $this->assertEquals(3.0, $product->getAverageRating());
         $this->assertNull($product->getGeolocation());
         $this->assertNull($product->getVideo());
         $this->assertCount(0, $product->getAttachments());
@@ -123,12 +123,14 @@ final class CatalogServiceTest extends ApiTestCase
         $this->assertSame(3, $companies[0]->getId());
         $this->assertSame('The World Company Inc.', $companies[0]->getName());
         $this->assertSame('the-world-company-inc', $companies[0]->getSlug());
-        $this->assertEquals(null, $companies[0]->getAverageRating());
+        $this->assertEquals(5.0, $companies[0]->getAverageRating());
         $this->assertNull($companies[0]->getImage());
         $this->assertTrue($companies[0]->isProfessional());
 
         $this->assertGreaterThanOrEqual(1400000000, $product->getCreatedAt()->getTimestamp());
         $this->assertGreaterThanOrEqual(0, $product->getCreatedAt()->diff($product->getUpdatedAt())->s);
+        $this->assertInstanceOf(\DateTimeInterface::class, $product->getAvailableSince());
+        $this->assertSame(1386316800, $product->getAvailableSince()->getTimestamp());
 
         $this->assertEmpty($product->getImages());
 
@@ -205,7 +207,7 @@ final class CatalogServiceTest extends ApiTestCase
         $this->assertTrue($product->isTransactional());
         $this->assertSame(0.0, $product->getGreenTax());
         $this->assertSame(1.23, $product->getWeight());
-        $this->assertNull($product->getAverageRating());
+        $this->assertEquals(3, $product->getAverageRating());
         $this->assertNull($product->getGeolocation());
         $this->assertNull($product->getVideo());
         $this->assertCount(0, $product->getAttachments());
@@ -217,13 +219,15 @@ final class CatalogServiceTest extends ApiTestCase
         $this->assertCount(1, $companies);
         $this->assertSame(3, $companies[0]->getId());
         $this->assertSame('The World Company Inc.', $companies[0]->getName());
-        $this->assertSame('the-world-company-inc.', $companies[0]->getSlug());
-        $this->assertNull($companies[0]->getAverageRating());
+        $this->assertSame('the-world-company-inc', $companies[0]->getSlug());
+        $this->assertEquals(5, $companies[0]->getAverageRating());
         $this->assertNull($companies[0]->getImage());
         $this->assertTrue($companies[0]->isProfessional());
 
         $this->assertGreaterThanOrEqual(1400000000, $product->getCreatedAt()->getTimestamp());
         $this->assertGreaterThanOrEqual(0, $product->getCreatedAt()->diff($product->getUpdatedAt())->s);
+        $this->assertSame(1386316800, $product->getAvailableSince()->getTimestamp());
+        $this->assertInstanceOf(\DateTimeInterface::class, $product->getAvailableSince());
 
         $this->assertEmpty($product->getImages());
     }
@@ -262,7 +266,7 @@ final class CatalogServiceTest extends ApiTestCase
         $this->assertTrue($product->isTransactional());
         $this->assertSame(0.0, $product->getGreenTax());
         $this->assertSame(1.23, $product->getWeight());
-        $this->assertNull($product->getAverageRating());
+        $this->assertEquals(3, $product->getAverageRating());
         $this->assertNull($product->getGeolocation());
         $this->assertNull($product->getVideo());
         $this->assertCount(0, $product->getAttachments());
@@ -274,13 +278,15 @@ final class CatalogServiceTest extends ApiTestCase
         $this->assertCount(1, $companies);
         $this->assertSame(3, $companies[0]->getId());
         $this->assertSame('The World Company Inc.', $companies[0]->getName());
-        $this->assertSame('the-world-company-inc.', $companies[0]->getSlug());
-        $this->assertNull($companies[0]->getAverageRating());
+        $this->assertSame('the-world-company-inc', $companies[0]->getSlug());
+        $this->assertEquals(5, $companies[0]->getAverageRating());
         $this->assertNull($companies[0]->getImage());
         $this->assertTrue($companies[0]->isProfessional());
 
         $this->assertGreaterThanOrEqual(1400000000, $product->getCreatedAt()->getTimestamp());
         $this->assertGreaterThanOrEqual(0, $product->getCreatedAt()->diff($product->getUpdatedAt())->s);
+        $this->assertInstanceOf(\DateTimeInterface::class, $product->getAvailableSince());
+        $this->assertSame(1386316800, $product->getAvailableSince()->getTimestamp());
 
         $this->assertEmpty($product->getImages());
     }