+++ tests/Pim/MultiVendorProduct/MultiVendorProductServiceTest.php
@@ -9,16 +9,21 @@ namespace Wizaplace\SDK\Tests\Pim\MultiVendorProduct;
 
 use Wizaplace\SDK\Exception\SomeParametersAreInvalid;
 use Wizaplace\SDK\File\File;
+use Wizaplace\SDK\Pagination;
 use Wizaplace\SDK\Pim\MultiVendorProduct\MultiVendorProduct;
 use Wizaplace\SDK\Pim\MultiVendorProduct\MultiVendorProductFile;
+use Wizaplace\SDK\Pim\MultiVendorProduct\MultiVendorProductFilter;
 use Wizaplace\SDK\Pim\MultiVendorProduct\MultiVendorProductService;
 use Wizaplace\SDK\Pim\MultiVendorProduct\MultiVendorProductStatus;
+use Wizaplace\SDK\Pim\MultiVendorProduct\MultiVendorProductVideo;
 use Wizaplace\SDK\Tests\ApiTestCase;
 use Wizaplace\SDK\Tests\File\Mock;
 
 final class MultiVendorProductServiceTest extends ApiTestCase
 {
     private $createdMvpId;
+
+    private const MVP_ID = 'a6e53f40-f4c5-3d56-af1d-cc83fd695feb';
 
     public function testGetMultiVendorProductById()
     {
@@ -74,6 +79,29 @@ final class MultiVendorProductServiceTest extends ApiTestCase
             2 => 5,
         ], $mvp->getAttributes());
         $this->assertSame([], $mvp->getImageIds());
+    }
+
+    public function testGetListMultiVendorProductByFilter()
+    {
+        $service = $this->buildMultiVendorProductService();
+
+        $filter = (new MultiVendorProductFilter())
+            ->setIds([self::MVP_ID, 'df8f3982-f9f0-49ec-a489-2903b4e3dace'])
+            ->setCodes(['Fire_trotti', 'EAN'])
+            ->setSupplierReferences(['REF_TROTTI'])
+            ;
+
+        $mvpList = $service->getListMultiVendorProduct($filter, 1, 2);
+
+        $this->assertInstanceOf(Pagination::class, $mvpList->getPagination());
+        $this->assertSame(5, $mvpList->getPagination()->getNbResults());
+        $this->assertSame(3, $mvpList->getPagination()->getNbPages());
+        $this->assertSame(1, $mvpList->getPagination()->getPage());
+        $this->assertCount(2, $mvpList->getMultiVendorProducts());
+
+        foreach ($mvpList->getMultiVendorProducts() as $mvp) {
+            $this->assertInstanceOf(MultiVendorProduct::class, $mvp);
+        }
     }
 
     public function testCreateMultiVendorProductWithInvalidPartialProduct()
@@ -225,6 +253,50 @@ final class MultiVendorProductServiceTest extends ApiTestCase
 
         $this->assertInstanceOf(MultiVendorProduct::class, $multiVendorProduct);
         $this->assertEquals($uuid, $multiVendorProduct->getId());
+    }
+
+    public function testAddVideoToMultiVendorProductWithHostedFile()
+    {
+        $service = $this->buildMultiVendorProductService();
+        $file = 'http://commondatastorage.googleapis.com/gtv-videos-bucket/sample/ForBiggerEscapes.mp4';
+        $multiVendorProductVideo = $service->addHostedVideoToMultiVendorProduct(static::MVP_ID, $file);
+
+        $this->assertInstanceOf(MultiVendorProductVideo::class, $multiVendorProductVideo);
+        $this->assertRegExp(
+            '/[0-9A-Fa-f]{8}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{12}/',
+            $multiVendorProductVideo->getId()
+        );
+        $this->assertNotEmpty($multiVendorProductVideo->getThumb());
+        $this->assertNotEmpty($multiVendorProductVideo->getPath());
+    }
+
+    public function testAddVideoToMultiVendorProductWithUploadedFile()
+    {
+        $service = $this->buildMultiVendorProductService();
+        $video = $this->mockUploadedFile("video.avi");
+        $file = new MultiVendorProductFile('file', $video->getStream(), $video->getClientFilename());
+
+        $multiVendorProductVideo = $service->addUploadedVideoToMultiVendorProduct(static::MVP_ID, $file);
+
+        $this->assertInstanceOf(MultiVendorProductVideo::class, $multiVendorProductVideo);
+        $this->assertRegExp(
+            '/[0-9A-Fa-f]{8}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{12}/',
+            $multiVendorProductVideo->getId()
+        );
+        $this->assertNotEmpty($multiVendorProductVideo->getThumb());
+        $this->assertNotEmpty($multiVendorProductVideo->getPath());
+    }
+
+    public function testDeleteVideoToMultiVendorProduct()
+    {
+        $service = $this->buildMultiVendorProductService();
+        $file = 'http://commondatastorage.googleapis.com/gtv-videos-bucket/sample/ForBiggerEscapes.mp4';
+        $service->addHostedVideoToMultiVendorProduct(static::MVP_ID, $file);
+
+        $service->deleteVideoToMultiVendorProduct(static::MVP_ID);
+        $mvp = $service->getMultiVendorProductById(static::MVP_ID);
+
+        $this->assertNull($mvp->getVideo()->getId());
     }
 
     private function buildMultiVendorProductService($userEmail = 'admin@wizaplace.com', $userPassword = 'password'): MultiVendorProductService